feat(#295) :: 원페이져 제출 메서드 controller매핑
