+++ src/main/java/team/jeonghokim/daedongyeojido/domain/onepager/presentation/OnePagerController.java
@@ -5,13 +5,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import team.jeonghokim.daedongyeojido.domain.onepager.presentation.dto.request.CommentRequest;
+import team.jeonghokim.daedongyeojido.domain.onepager.presentation.dto.request.SubmitOnePagerRequest;
 import team.jeonghokim.daedongyeojido.domain.onepager.service.CreateRejectedOnePagerCommentService;
+import team.jeonghokim.daedongyeojido.domain.onepager.service.CreateSubmitOnePagerService;
 
 @RestController
 @RequestMapping("/onepager")
 @RequiredArgsConstructor
 public class OnePagerController {
     private final CreateRejectedOnePagerCommentService createRejectedOnePagerCommentService;
+    private final CreateSubmitOnePagerService createSubmitOnePagerService;
 
     @PostMapping("/submissions/{submission-id}/comment")
     @ResponseStatus(HttpStatus.CREATED)
@@ -20,5 +23,11 @@ public class OnePagerController {
         @PathVariable("submission-id") Long submissionId
     ) {
         createRejectedOnePagerCommentService.execute(request, submissionId);
+    }
+
+    @PostMapping("/{form-id}/submissions")
+    @ResponseStatus(HttpStatus.CREATED)
+    public void createSubmission(@RequestBody @Valid SubmitOnePagerRequest request) {
+        createSubmitOnePagerService.execute(request);
     }
 }
